I forgot to update the VS2005 and Code::Blocks projects in last commit :)
git-svn-id: https://sfml.svn.sourceforge.net/svnroot/sfml/branches/sfml2@1418 4e206d99-4929-0410-ac5d-dfc041789085

+++ src/SFML/Window/Win32/WindowImplWin32.cpp
@@ -674,7 +674,7 @@ Key::Code WindowImplWin32::VirtualKeyCodeToSF(WPARAM key, LPARAM flags)
         case VK_SHIFT :
         {
             static UINT lShift = MapVirtualKey(VK_LSHIFT, MAPVK_VK_TO_VSC);
-            UINT scancode = (flags & (0xFF << 16)) >> 16;
+            UINT scancode = static_cast<UINT>((flags & (0xFF << 16)) >> 16);
             return scancode == lShift ? Key::LShift : Key::RShift;
         }
